Add comments and fix physics world

+++ include/Math/Constants.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-inline constexpr float PI = static_cast<float>(3.14159265358979323846264338327950);
-inline constexpr float PI_TWO = static_cast<float>(6.283185307179586);
+inline constexpr double PI = static_cast<double>(3.14159265358979323846264338327950);
+inline constexpr double PI_TWO = static_cast<double>(6.283185307179586);
 
-inline constexpr float DEG_TO_RAD = static_cast<float>(0.01745329251994329576923690768489);
-inline constexpr float RAD_TO_DEG = static_cast<float>(57.295779513082320876798154814105);
+inline constexpr double DEG_TO_RAD = static_cast<double>(0.01745329251994329576923690768489);
+inline constexpr double RAD_TO_DEG = static_cast<double>(57.295779513082320876798154814105);
 
-inline constexpr float Sqrt2 = static_cast<float>(1.41421356237309504880168);
+inline constexpr double Sqrt2 = static_cast<double>(1.41421356237309504880168);
